requirements.txt
flask==2.3.3
mysql-connector-python==8.1.0
requests==2.31.0
pyyaml==6.0.1
python-dateutil==2.8.2